remove __LINUX__

+++ packages/xunsearch-1.4.9/src/task.cc
@@ -1957,7 +1957,7 @@ static int zcmd_scws_get(XS_CONN *conn)
 		
 		KeyExtract_Init(NULL, UTF8_CODE);
 		string bufstr(XS_CMD_BUF(cmd), XS_CMD_BLEN(cmd));
-		const char *keywords_str = KeyExtract_GetKeyWords(bufstr.c_str(), cmd->arg2, true);
+		const char *keywords_str = KeyExtract_GetKeyWords(bufstr.c_str(), (int)cmd->arg2, true);
 		std::vector<string> keywords;
 		splitKeyWords(keywords_str, keywords, "#");
 		for(int i = 0; i < keywords.size(); i++)
